Fix the icon in the Dock for Mac (#505)
There is a strange line in `texmacs.cpp` which reset the icon. I've now
fixed it to point to the new icon. I do not understand why this is
necessary. I guess we can remove that line since the application icon
should be specified in the metadata and not in the code (unless one what
to change it programmatically)