[11_5] Font: use fontconfig on Linux to detect font paths
Currently a pre-defined list of paths is searched for fonts. On my system, there are no fonts in the directories that are searched ("/usr/share/fonts/opentype" and "/usr/share/fonts/truetype"), so no additional fonts are found.

On Linux, fontconfig is widely used to specify the locations of fonts on the system. The PR uses fontconfig to get a list of font directories and search them for fonts.

Co-authored-by: Jeroen Wouters <jeroen@diasporist.org>
Reviewed-on: https://codeberg.org/XmacsLabs/mogan/pulls/33
Co-authored-by: woutersj <woutersj@noreply.codeberg.org>
Co-committed-by: woutersj <woutersj@noreply.codeberg.org>