[23_27] Draw an arc by center and two points (#1038)
## Why you open this Pull Request?
Draw an arc by center and two points.


Enter three points. The first is the center of the circle and the second
is the starting point of the arc. Draw a straight line through the third
point and the center of the circle. The intersection of the straight
line and the circle is the end point of the arc. Use clockwise method.



## What work have you done in the current Pull Request?
- [x] Draw an arc by center and two points.