[75_12] Fix rendering of linked image on Windows using absolute path
## What and Why
It was a fix on branch `lts-125`.

For `branch-1.2`, since we have upgraded to lolly 1.4.x, the implicit
conversion from string to url is using url_system now. This pull request
is for using url_system explicitly but not a fix now. Let us keep the
pull request title in `branch-1.2`.

## How to test your changes?
+ [x] Linux
+ [x] Windows
+ [ ] macOS

Insert a linked image with absolute path to see if it is rendered
correctly.

Co-authored-by: 沈浪熊猫儿 <shenda@ustc.edu>