[11_31] Import font to $TEXMACS_HOME_PATH/fonts/truetype
## What
+ Import the font to $TEXMACS_HOME_PATH/fonts/truetype
  - Use the md5 digest as the font file name
- No longer loading fonts from the font local path on
macOS/Linux/Windows
+ Collect the font info when loading the font database
+ Fix the bug of tt_fint_font when the font file is using TTF/TTC/OTC as
the suffix

TODO: in later pull request: the dialogue to import font should only
allow ttf/ttc/otc

## Why
`ft_load_face` does not work on windows when the font path contains
Chinese characters (eg. `方正小标宋简.ttf`). To avoid Chinese font path name,
we use the content md5 digest as the file name when import the font.

The goal of this pull request is to support `方正小标宋简.ttf`, and because of
the bug of ft_load_face, we have to re-impl the mechanism of font
importing.

## How to test your changes?
+ [x] Linux
+ [ ] macOS
+ [x] Windows

1. Open `Document -> Fonts` and click `Import...` to Import font.
2. Click `Tools -> Fonts -> Clear font cache`, and restart Mogan, the
imported fonts should still work