[81_1] Add a blank tab page toolbar to the main window (default invisible currently)
Add a blank tab page toolbar to the main window. Due to the current
blank toolbar, it is not displayed by default. You can display it by the
main menu "view" -> "Tab bar".
I have **modified** 11 files and **added** 2 new files. I understand
that the recommended practice is to have small code merge requests.
Although I have modified many files, most of these modifications are
only a few lines of code and are necessary.

For example, the changes in the following 7 files are all within **3
lines**:

- TeXmacs/progs/init-research.scm
- TeXmacs/progs/texmacs/menus/view-menu.scm
- src/Edit/Interface/edit_interface.cpp
- src/Graphics/Gui/widget.cpp
- src/Texmacs/Data/new_window.cpp
- src/Texmacs/Window/tm_frame.cpp
- src/Texmacs/Window/tm_window.cpp

Why do so many files need to be modified even if the changes are small?
Because this follows the existing practice of how the main toolbar is
implemented, and I need to follow the same process.

Additionally, the two newly added files,

- TeXmacs/progs/texmacs/menus/tabpage-menu.scm
- src/Plugins/Qt/QTMTabPage.hpp

are actually placeholder files intended to lay the groundwork for
subsequent code.

Therefore, excluding these files, the main file I modified is
`qt_tm_widget.cpp`.

Build and run, you can display it by the main menu "view" -> "Tab bar".