[70_1] Added dependencies for tree-sitter and tree-sitter-cpp (#1940)
<!-- Thank you for your contribution! -->
## What
Added dependencies on packages **tree-sitter** and **tree-sitter-cpp**
## Why
Syntax highlighting depends on tree-sitter and the corresponding
language rules.
Tree-sitter has an existing script on xmake, and the script for
tree-sitter-cpp is located in xmake\packages\t\tree-sitter-cpp.
The tests for tree-sitter-cpp depend on tree-sitter, so on_test is not
written separately. Subsequent development will show that
tree-sitter-cpp can be imported normally.
## How to test your changes?
```
xmake config -v
```
or
```
xmake build research
```
Both packages can be correctly downloaded and installed. The on_test for
tree-sitter can pass during this process.

---------

Co-authored-by: Your Name <9838919+unbsky@user.noreply.gitee.com>