[24_20] Fix the problem of missing titles and images when exporting to docx (#2066)
<!-- Thank you for your contribution! -->
## What
Fix the problem of missing titles and images when exporting to docx.

## Why
In previous versions, exporting `tm`, `tmu` files to `docx` format would
be missing titles and images. The main causes are as follows:
1. Missing title: The problem arises because string-save saves the
output of `serialize-html` function, which only converts `html-stree` to
`html-snippet` and is not the final output of exporting to html. Change
it to `export-buffer-main` to include the export to html process
completely;
2. Missing pictures: the missing pictures are mainly due to the fact
that when executing the pandoc command, the execution environment is not
consistent with the path of exporting to html, so the picture
information can not be obtained, which has been fixed now.

## How to test your changes?
tm：

![图片](https://github.com/user-attachments/assets/c282249d-332d-468c-93d6-c98de1e0e180)

html:

![图片](https://github.com/user-attachments/assets/f58d12b9-bd16-463c-9fd3-3c988bcde955)

docx:

![图片](https://github.com/user-attachments/assets/2b2fc234-2bc9-496d-8965-ce211c9e3f4c)