[improve] Add M-a/C-a event for QTMLineEdit (#2233)
<!-- Thank you for your contribution! -->

## What  
Added support for `Cmd+A` / `Ctrl+A` key events in `QTMLineEdit` to
select all text in the search toolbar.

Related: #2232

## Why  
`Cmd+A` / `Ctrl+A` is a native `QLineEdit` behavior in Qt. Supporting it
improves usability by allowing users to quickly modify their search
query. This standard interaction should not be ignored, I guess.

## How to test your changes  
1. Build the project.  
2. Open the search toolbar using `Cmd+f` / `Ctrl+f`.  
3. Enter any text.  
4. Press `Cmd+a` / `Ctrl+a` to ensure all text is selected.